Changed reservation id to long, changed ReservationRepository to ReservationFileRepository, extracted interface ReservationRepository, added new class ReservationDatabaseRepository and implemented CRUD.

+++ src/main/java/org/example/domain/guest/GuestService.java
@@ -52,7 +52,7 @@ public class GuestService {
         repository.edit(id, firstName, lastName, age, gender);
     }
 
-    public Guest getGuestById(int Id) {
+    public Guest getGuestById(long Id) {
         return repository.findById(Id);
     }
