+++ src/main/java/org/example/ui/gui/reservations/AddNewReservationScene.java
@@ -46,13 +46,13 @@ public class AddNewReservationScene {
         List<RoomDTO> allAsDTO = this.roomService.getAllAsDTO();
         List<RoomSelectionItem> roomSelectionItems = new ArrayList<>();
         allAsDTO.forEach(dto -> {
-            roomSelectionItems.add(new RoomSelectionItem(dto.getNumber(), (int)dto.getId()));
+            roomSelectionItems.add(new RoomSelectionItem(dto.getNumber(), (int) dto.getId()));
         });
 
         List<GuestDTO> guestsAsDTO = this.guestService.getGuestsAsDTO();
         List<GuestSelectionItem> guestSelectionItems = new ArrayList<>();
         guestsAsDTO.forEach(dto -> {
-            guestSelectionItems.add(new GuestSelectionItem(dto.getFirstName(), dto.getLastName(), (int)dto.getId()));
+            guestSelectionItems.add(new GuestSelectionItem(dto.getFirstName(), dto.getLastName(), (int) dto.getId()));
         });
 
         Label roomLabel = new Label("Pokój:");
@@ -75,7 +75,7 @@ public class AddNewReservationScene {
             LocalDate from = fromDateField.getValue();
             LocalDate to = toDateField.getValue();
             int guestId = guestField.getValue().getId();
-            int roomId = roomField.getValue().getId();
+            long roomId = roomField.getValue().getId();
 
             try {
                 this.reservationService.createNewReservation(from, to, roomId, guestId);
